Update Desmos API Key
The API Key is in the source code with permission from Desmos Studio PBC

Here is their exact response to my email regarding compliance with Desmos API TOS part 5C, if anyone is concerned:

For folks looking at your source code or on Github, we recommend that you add a comment that says something like the below (note that you'll need to replace the highlighted text below with the name of your project):

<!-- This API key is only allowed for use on [name of your project]. If you'd like to add Desmos to your fork of [name of your project] or for use in your own project, please visit www.desmos.com/my-api or reach out to the Desmos Studio team at partnerships@desmos.com -->

And appreciate the question. Over time we might add domain-level restrictions by key, but we haven't done that yet, so for now we're largely relying on good faith usage and including messages like the above in open source projects.